requirements.txt
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu126

accelerate>=1.13.0
datasets>=3.0.0
duckdb>=1.1.0
fastapi>=0.136.0
huggingface-hub
lark>=1.3.1
llama-cpp-python
outlines>=1.2.12
polars>=1.20.0
pydantic>=2.13.2
pygbnf
torch>=2.11.0
transformers>=5.5.4
uvicorn[standard]>=0.30.0
vllm
xgrammar>=0.1.33
